src/requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
boto3==1.22.7
botocore==1.25.7
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.3
django-autoslug==1.9.8
django-colorfield==0.6.3
django-richtextfield==1.6.1
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
jmespath==1.0.0
kappa==0.6.0
Pillow==9.1.0
placebo==0.9.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.0
troposphere==4.0.1
urllib3==1.26.9
Werkzeug==2.1.2
wsgi-request-logger==0.4.6
zappa==0.54.1
